refactor: move twofish into package
