+++ packages/twofish/src/c/string.h
