--- src/core/encryption/ciphers/Twofish/wasm/README.md
@@ -1,4 +1,0 @@
-The twofish code is taken from https://sources.debian.org/src/twofish/0.3-5
-
-TODO:
-- Add tests for WASM module against reference vectors
